+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.2
 
 - Added example project
-- Added implicit factory for `RePaint` widget
+- Added implicit inline factory for `RePaint` widget
 
 ## 0.0.1